refactor: extract Quarkus plugin tests to separate file
Move the Quarkus-specific tests from PluginUpgradeStrategyTest to a new
PluginUpgradeQuarkusTest class to stay well within the 2000-line
Checkstyle limit and leave room for future plugin additions.
